+++ beets/util/__init__.py
@@ -16,6 +16,7 @@
 
 from __future__ import (division, absolute_import, print_function,
                         unicode_literals)
+import ctypes
 
 import os
 import sys
@@ -760,3 +761,52 @@ def interactive_open(targets, command=None):
     command += targets
 
     return os.execlp(*command)
+
+
+def is_filesystem_case_sensitive(path):
+    """Checks if the filesystem at the given path is case sensitive.
+    If the path does not exist, a case sensitive file system is
+    assumed if the system is not windows.
+
+    :param path: The path to check for case sensitivity.
+    :return: True if the file system is case sensitive, False else.
+    """
+    if os.path.exists(path):
+        # Check if the path to the library exists in lower and upper case
+        if os.path.exists(path.lower()) and \
+           os.path.exists(path.upper()):
+            # All the paths may exist on the file system. Check if they
+            # refer to different files
+            if platform.system() != 'Windows':
+                # os.path.samefile is only available on Unix systems for
+                # python < 3.0
+                return not os.path.samefile(path.lower(),
+                                            path.upper())
+
+            # On windows we use GetLongPathNameW to determine the real path
+            # using the actual case.
+            def get_long_path_name(short_path):
+                if not isinstance(short_path, unicode):
+                    short_path = unicode(short_path)
+                buf = ctypes.create_unicode_buffer(260)
+                get_long_path_name_w = ctypes.windll.kernel32.GetLongPathNameW
+                return_value = get_long_path_name_w(short_path, buf, 260)
+                if return_value == 0 or return_value > 260:
+                    # An error occurred
+                    return short_path
+                else:
+                    long_path = buf.value
+                    # GetLongPathNameW does not change the case of the drive
+                    # letter.
+                    if len(long_path) > 1 and long_path[1] == ':':
+                        long_path = long_path[0].upper() + long_path[1:]
+                    return long_path
+
+            lower = get_long_path_name(path.lower())
+            upper = get_long_path_name(path.upper())
+
+            return lower != upper
+        else:
+            return True
+    # By default, the case sensitivity depends on the platform.
+    return platform.system() != 'Windows'